// src/types/types.ts
export interface JsonObject {
  [name: string]: any // TODO: this needs to become `unknown`
}

/** A collection of unknown extra methods exposed by a plugin. */
export interface EdgeOtherMethods {
  readonly [name: string]: any
}

/** We frequently index things by pluginId, so provide a helper. */
export interface EdgePluginMap<Value> {
  [pluginName: string]: Value
}
// ...
export interface EdgeIo {
  // Crypto:
  readonly random: EdgeRandomFunction
  readonly scrypt: EdgeScryptFunction

  // Local io:
  readonly disklet: Disklet
  readonly fetch: EdgeFetchFunction

  // This is only present if the platform has some way to avoid CORS:
  readonly fetchCors?: EdgeFetchFunction

  // Deprecated:
  // eslint-disable-next-line no-use-before-define
  readonly console: EdgeConsole
  readonly WebSocket: EdgeWebSocket
}
// ...
export interface EdgeNativeIo {
  [packageName: string]: EdgeOtherMethods
}
// ...
export interface EdgeWalletInfo {
  id: string
  type: string
  keys: JsonObject
}

export type EdgeWalletInfoFull = EdgeWalletInfo & {
  appIds: string[]
  archived: boolean
  deleted: boolean
  hidden: boolean
  sortIndex: number
}

export interface EdgeWalletState {
  archived?: boolean
  deleted?: boolean
  hidden?: boolean
  sortIndex?: number
}

export interface EdgeWalletStates {
  [walletId: string]: EdgeWalletState
}
// ...
export interface EdgeCurrencyWallet {
  readonly on: Subscriber<EdgeCurrencyWalletEvents>
  readonly watch: Subscriber<EdgeCurrencyWallet>

  // Data store:
  readonly id: string
  readonly keys: JsonObject
  readonly type: string
  readonly publicWalletInfo: EdgeWalletInfo
  readonly disklet: Disklet
  readonly localDisklet: Disklet
  sync(): Promise<void>

  // Wallet keys:
  readonly displayPrivateSeed: string | null
  readonly displayPublicSeed: string | null

  // Wallet name:
  readonly name: string | null
  renameWallet(name: string): Promise<void>

  // Fiat currency option:
  readonly fiatCurrencyCode: string
  setFiatCurrencyCode(fiatCurrencyCode: string): Promise<void>

  // Currency info:
  readonly currencyInfo: EdgeCurrencyInfo
  nativeToDenomination(
    nativeAmount: string,
    currencyCode: string
  ): Promise<string>
  denominationToNative(
    denominatedAmount: string,
    currencyCode: string
  ): Promise<string>

  // Chain state:
  readonly balances: EdgeBalances
  readonly blockHeight: number
  readonly syncRatio: number

  // Running state:
  startEngine(): Promise<void>
  stopEngine(): Promise<void>

  // Token management:
  changeEnabledTokens(currencyCodes: string[]): Promise<void>
  enableTokens(tokens: string[]): Promise<void>
  disableTokens(tokens: string[]): Promise<void>
  getEnabledTokens(): Promise<string[]>
  addCustomToken(token: EdgeTokenInfo): Promise<void>

  // Transaction history:
  getNumTransactions(opts?: EdgeCurrencyCodeOptions): Promise<number>
  getTransactions(opts?: EdgeGetTransactionsOptions): Promise<EdgeTransaction[]>

  // Addresses:
  getReceiveAddress(opts?: EdgeCurrencyCodeOptions): Promise<EdgeReceiveAddress>
  saveReceiveAddress(receiveAddress: EdgeReceiveAddress): Promise<void>
  lockReceiveAddress(receiveAddress: EdgeReceiveAddress): Promise<void>

  // Sending:
  makeSpend(spendInfo: EdgeSpendInfo): Promise<EdgeTransaction>
  signTx(tx: EdgeTransaction): Promise<EdgeTransaction>
  broadcastTx(tx: EdgeTransaction): Promise<EdgeTransaction>
  saveTx(tx: EdgeTransaction): Promise<void>
  sweepPrivateKeys(edgeSpendInfo: EdgeSpendInfo): Promise<EdgeTransaction>
  saveTxMetadata(
    txid: string,
    currencyCode: string,
    metadata: EdgeMetadata
  ): Promise<void>
  getMaxSpendable(spendInfo: EdgeSpendInfo): Promise<string>
  getPaymentProtocolInfo(
    paymentProtocolUrl: string
  ): Promise<EdgePaymentProtocolInfo>

  // Wallet management:
  resyncBlockchain(): Promise<void>
  dumpData(): Promise<EdgeDataDump>
  getDisplayPrivateSeed(): string | null
  getDisplayPublicSeed(): string | null

  // Data exports:
  exportTransactionsToQBO(opts: EdgeGetTransactionsOptions): Promise<string>
  exportTransactionsToCSV(opts: EdgeGetTransactionsOptions): Promise<string>

  // URI handling:
  parseUri(uri: string, currencyCode?: string): Promise<EdgeParsedUri>
  encodeUri(obj: EdgeEncodeUri): Promise<string>

  readonly otherMethods: EdgeOtherMethods

  // Deprecated API's:
  getBalance(opts?: EdgeCurrencyCodeOptions): string
  getBlockHeight(): number
}
// ...
export interface EdgeCreateCurrencyWalletOptions {
  fiatCurrencyCode?: string
  name?: string

  // Create a private key from some text:
  importText?: string

  // Used to tell the currency plugin what keys to create:
  keyOptions?: JsonObject

  // Used to copy wallet keys between accounts:
  keys?: JsonObject
}

export interface EdgeCurrencyConfig {
  readonly watch: Subscriber<EdgeCurrencyConfig>

  readonly currencyInfo: EdgeCurrencyInfo
  readonly otherMethods: EdgeOtherMethods
  readonly userSettings: JsonObject | undefined

  changeUserSettings(settings: JsonObject): Promise<void>
  importKey(userInput: string): Promise<JsonObject>
}

export interface EthereumTransaction {
  chainId: number // Not part of raw data, but needed for signing
  nonce: string
  gasPrice: string
  gasLimit: string
  to: string
  value: string
  data: string
  // The transaction is unsigned, so these are not present:
  v?: string
  r?: string
  s?: string
}

// rates ---------------------------------------------------------------

export interface EdgeRateCacheEvents {
  close: void
  update: unknown
}

export interface EdgeRateCache {
  readonly on: Subscriber<EdgeRateCacheEvents>

  convertCurrency(
    fromCurrency: string,
    toCurrency: string,
    amount: number
  ): Promise<number>
}

// swap ----------------------------------------------------------------

/**
 * Information and settings for a currency swap plugin.
 */
export interface EdgeSwapConfig {
  readonly watch: Subscriber<EdgeSwapConfig>

  readonly enabled: boolean
  readonly needsActivation: boolean
  readonly swapInfo: EdgeSwapInfo
  readonly userSettings: JsonObject | undefined

  changeEnabled(enabled: boolean): Promise<void>
  changeUserSettings(settings: JsonObject): Promise<void>
}

export type EdgeSwapQuote = EdgeSwapPluginQuote & {
  readonly isEstimate: boolean // No longer optional at this point
  readonly quoteUri?: string
}

// edge login ----------------------------------------------------------

export interface EdgeLoginRequest {
  readonly appId: string
  approve(): Promise<void>

  readonly displayName: string
  readonly displayImageUrl: string | undefined
}

export interface EdgeLobby {
  readonly loginRequest: EdgeLoginRequest | undefined
  // walletRequest: EdgeWalletRequest | undefined
}

// storage -------------------------------------------------------------

export interface EdgeDataStore {
  deleteItem(storeId: string, itemId: string): Promise<void>
  deleteStore(storeId: string): Promise<void>

  listItemIds(storeId: string): Promise<string[]>
  listStoreIds(): Promise<string[]>

  getItem(storeId: string, itemId: string): Promise<string>
  setItem(storeId: string, itemId: string, value: string): Promise<void>
}

// account -------------------------------------------------------------

export interface EdgeAccountEvents {
  close: void
}

export interface EdgeAccount {
  readonly on: Subscriber<EdgeAccountEvents>
  readonly watch: Subscriber<EdgeAccount>

  // Data store:
  readonly id: string
  readonly keys: JsonObject
  readonly type: string
  readonly disklet: Disklet
  readonly localDisklet: Disklet
  sync(): Promise<void>

  // Basic login information:
  readonly appId: string
  readonly loggedIn: boolean
  readonly loginKey: string
  readonly recoveryKey: string | undefined // For email backup
  readonly username: string

  // Special-purpose API's:
  readonly currencyConfig: EdgePluginMap<EdgeCurrencyConfig>
  readonly rateCache: EdgeRateCache
  readonly swapConfig: EdgePluginMap<EdgeSwapConfig>
  readonly dataStore: EdgeDataStore

  // What login method was used?
  readonly edgeLogin: boolean
  readonly keyLogin: boolean
  readonly newAccount: boolean
  readonly passwordLogin: boolean
  readonly pinLogin: boolean
  readonly recoveryLogin: boolean

  // Change or create credentials:
  changePassword(password: string): Promise<void>
  changePin(opts: {
    pin?: string // We keep the existing PIN if unspecified
    enableLogin?: boolean // We default to true if unspecified
  }): Promise<string>
  changeRecovery(questions: string[], answers: string[]): Promise<string>

  // Verify existing credentials:
  checkPassword(password: string): Promise<boolean>
  checkPin(pin: string): Promise<boolean>

  // Remove credentials:
  deletePassword(): Promise<void>
  deletePin(): Promise<void>
  deleteRecovery(): Promise<void>

  // OTP:
  readonly otpKey: string | undefined // OTP is enabled if this exists
  readonly otpResetDate: string | undefined // A reset is requested if this exists
  cancelOtpReset(): Promise<void>
  disableOtp(): Promise<void>
  enableOtp(timeout?: number): Promise<void>

  // Edge login approval:
  fetchLobby(lobbyId: string): Promise<EdgeLobby>

  // Login management:
  logout(): Promise<void>

  // Master wallet list:
  readonly allKeys: EdgeWalletInfoFull[]
  changeWalletStates(walletStates: EdgeWalletStates): Promise<void>
  createWallet(type: string, keys?: JsonObject): Promise<string>
  getFirstWalletInfo(type: string): EdgeWalletInfo | undefined
  getWalletInfo(id: string): EdgeWalletInfo | undefined
  listWalletIds(): string[]
  listSplittableWalletTypes(walletId: string): Promise<string[]>
  splitWalletInfo(walletId: string, newWalletType: string): Promise<string>

  // Currency wallets:
  readonly activeWalletIds: string[]
  readonly archivedWalletIds: string[]
  readonly hiddenWalletIds: string[]
  readonly currencyWallets: { [walletId: string]: EdgeCurrencyWallet }
  createCurrencyWallet(
    type: string,
    opts?: EdgeCreateCurrencyWalletOptions
  ): Promise<EdgeCurrencyWallet>
  waitForCurrencyWallet(walletId: string): Promise<EdgeCurrencyWallet>

  // Web compatibility:
  signEthereumTransaction(
    walletId: string,
    transaction: EthereumTransaction
  ): Promise<string>

  // Swapping:
  fetchSwapQuote(request: EdgeSwapRequest): Promise<EdgeSwapQuote>

  // Deprecated names:
  // eslint-disable-next-line no-use-before-define
  readonly pluginData: EdgePluginData
  readonly exchangeCache: EdgeRateCache
  readonly currencyTools: EdgePluginMap<EdgeCurrencyConfig>
  readonly exchangeTools: EdgePluginMap<EdgeSwapConfig>
  getExchangeQuote(request: EdgeSwapRequest): Promise<EdgeSwapQuote>
}
// ...
export interface EdgeContextOptions {
  apiKey: string
  appId: string
  authServer?: string
  hideKeys?: boolean
  path?: string // Only used on node.js
  plugins?: EdgeCorePluginsInit
}
// ...
export interface EdgeContext {
  readonly on: Subscriber<EdgeContextEvents>
  readonly watch: Subscriber<EdgeContext>
  close(): Promise<void>

  readonly appId: string

  // Local user management:
  localUsers: EdgeUserInfo[]
  fixUsername(username: string): string
  listUsernames(): Promise<string[]>
  deleteLocalAccount(username: string): Promise<void>

  // Account creation:
  usernameAvailable(username: string): Promise<boolean>
  createAccount(
    username: string,
    password?: string,
    pin?: string,
    opts?: EdgeAccountOptions
  ): Promise<EdgeAccount>

  // Edge login:
  requestEdgeLogin(opts: EdgeEdgeLoginOptions): Promise<EdgePendingEdgeLogin>

  // Fingerprint login:
  loginWithKey(
    username: string,
    loginKey: string,
    opts?: EdgeAccountOptions
  ): Promise<EdgeAccount>

  // Password login:
  checkPasswordRules(password: string): EdgePasswordRules
  loginWithPassword(
    username: string,
    password: string,
    opts?: EdgeAccountOptions
  ): Promise<EdgeAccount>

  // PIN login:
  pinLoginEnabled(username: string): Promise<boolean>
  loginWithPIN(
    username: string,
    pin: string,
    opts?: EdgeAccountOptions
  ): Promise<EdgeAccount>

  // Recovery2 login:
  getRecovery2Key(username: string): Promise<string>
  loginWithRecovery2(
    recovery2Key: string,
    username: string,
    answers: string[],
    opts?: EdgeAccountOptions
  ): Promise<EdgeAccount>
  fetchRecovery2Questions(
    recovery2Key: string,
    username: string
  ): Promise<string[]>
  listRecoveryQuestionChoices(): Promise<string[]>

  // OTP stuff:
  requestOtpReset(username: string, otpResetToken: string): Promise<Date>
  fetchLoginMessages(): Promise<EdgeLoginMessages>

  // Background mode:
  readonly paused: boolean
  changePaused(paused: boolean, opts?: { secondsDelay?: number }): Promise<void>

  // Deprecated API's:
  pinExists(username: string): Promise<boolean>
}

// ---------------------------------------------------------------------
// fake mode
// ---------------------------------------------------------------------

export interface EdgeFakeUser {
  username: string
  loginId: string
  loginKey: string
  repos: { [repo: string]: { [path: string]: JsonObject } }
  server: JsonObject
}

export interface EdgeFakeWorld {
  close(): Promise<void>

  makeEdgeContext(
    opts: EdgeContextOptions & { cleanDevice?: boolean }
  ): Promise<EdgeContext>

  goOffline(offline?: boolean): Promise<void>
  dumpFakeUser(account: EdgeAccount): Promise<EdgeFakeUser>
}